Adding the file 7-print_last_digit

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -13,6 +13,6 @@ int print_last_digit(int num)
 
 	_putchar('0' + r);
 
-	return (r);
+	return ((int)(r));
 }
